roma/requirements.txt
langgraph>=0.0.55
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
pathlib>=1.0.1
chardet>=5.2.0
python-magic>=0.4.27
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.5.0
pyyaml>=6.0.1
requests>=2.31.0
aiofiles>=23.2.0
asyncio>=3.4.3
concurrent-futures>=3.1.1